+++ build.gradle
@@ -69,6 +69,7 @@ dependencies {
   provided 'javax.servlet:javax.servlet-api:3.1.0'
 
   testCompile "org.grails:grails-plugin-testing"
+  testCompile "com.icegreen:greenmail:1.4.0"
 
   compile "javax.mail:javax.mail-api:1.5.1"
   compile "com.sun.mail:javax.mail:1.5.1"
